RPi.GPIO compatibility: Constant value fixes
Use RPi.GPIO-compatible values of BOARD and BCM.

gpio_function() is supposed to return an integer, not a string. Fix the
values of IN, OUT, and _UNEXPORTED to solve this. Rename _UNEXPORTED to
UNKNOWN to match RPi.GPIO.

Define pull and edge constants in terms of their offset values to prevent
inconsistency, just like RPi.GPIO. Fix the value of _EDGE_OFFSET to match
RPi.GPIO.