Cleanups to cleanup()
Lift error-checking out of _cleanup_one(); make the user API cleanup()
skip cleaning up any invalid channels.

Don't reconfigure a GPIO as input before unexporting it. It's up to the
kernel to decide what to do when unexporting a GPIO, so avoid user-space
operations that might cause additional transitions on the pin. This also
paves the way for PWM pin usage in the future; in this case, setting a
pin to GPIO input on clean up doesn't make any sense.

Only clean up channels that the library exported, to avoid interfering
with any channels exported by other applications.