Identify boards using DT compatible value
Some people consider DT's model property not to be part of the ABI, and
the values it contains are likely to change over time since it's somewhat
free-form marketing-oriented data. Instead, use the compatible property
to identify boards since this is a more stable engineering-oriented
value.

Fix identification of Jetson AGX Xavier so that it actually works; the
code was previously looking for the wrong model value.

Add support to identify Jetson TX2i, a/k/a "storm".

Fix the code that determines the revision of Jetson Nano so that it
explicitly looks at the module revision rather than the revision of
whatever file/property just happened to be returned first by
os.listdir().