Warn if running on an unsupported carrier board
It is likely that any carrier board other than the developer kit carrier
board uses a different set of GPIOs for any GPIO expansion connector that
may be present. Thus, it is unlikely the code will operate correctly with
any non-default carrier board. Warn when this situation is detected, or
if the files necessary to determine this are not present. This notifies
the user of the issue, without preventing use of the library in the case
where the carrier board does happen to use the expected set of GPIOs.