Add test_all_pins_input.py sample
This reads the input value of every implemented pin and prints it. It's a
quick way to check that the current board is identified and that all
expected pins are implemented, in a way that's suitable for use regardless
of what hardware may or may not be attached to the GPIO expansion header.
This script will be used in NVIDIA's internal test system.