PWM: Set initial duty cycle to 0%
The sysfs API only allows setting duty cycle and period separately, not
atomically. The PWM core enforces that the duty time is less than the period,
even when the PWM channel is disabled. The only way to make all this work
together when the PWM is initially configured from a completely unknown state
is to set the initial duty cycle to 0%, and do so before changing the
frequency. RPi.GPIO initially selects a 0% duty cycle when initializing a PWM
channel, and the start() API has always required the user to specify the
desired duty cycle, so this change shouldn't cause any compatibility issues.