Fix event debounce thread to always unlock mutex
When a "bouncetime" value was set, and an callback was skipped because of
this, the mutex wasn't released, which prevented the thread from making
further progress. Fix this by using a finally block to guarantee that the
mutex is always released.

Add some samples that allow the problem to be easily reproduced.

Fixes github.com issue #40