Warn when pull_up_down argument is passed to setup()
Jetson.GPIO doesn't implement the pull_up_down argument to setup().
Rather, pin pulls must be configured via the Tegra pinmux, e.g. using the
Jetson-IO tool or pinmux spreadsheet. Warn the user if the argument is
explicitly passed. The argument exists to support API compatibility with
RPi.GPIO.

Note: The warning is emitted any time this argument is passed, rather than
when an argument value is passed that doesn't match the pin's
configuration; there is no way for the code to determine the current
configuration to compare it against the passed value.