PWM: cache sysfs files for perf
The existing code opens, writes, and closes a sysfs file every time a
PWM's duty value is changed. This takes limits the maximum rate at which
the duty cycle can be changed. Enhance the code to open the file whenever
a PWM is exported, so that updating the duty cycle is implemented simply
by seeking/writing to an existing file handle.