PWM: don't stop on pure duty cycle changes
When changing just the PWM duty cycle, there is no need to juggle the
order of changing period and duty cycle values to avoid temporarily
out-of-range configurations. Consequently, we do not need to stop the PWM
running in this case. Enhance the code to only stop the PWM, or set the
period value, when the frequency changes or when the PWM is first
initialized. This avoids glitches in the PWM output when changing just the
duty cycle.